feat: Add check for metered connections
When running an update, `uupd` does not check if the user is currently on a metered connection (e.g.: routing cellular network). In combination with automatic background updates, this may
result in unwanted charges being applied.

This feature adds a check for metered connections by querying the NetworkManager dbus interface. If the connection is metered, the update will be aborted (checking for updates still happens).
The user can manually override this behaviour by supplying the `-ignore-metered-connection` flag to `uupd`.